Clean up AbilitySystemSetupComponent a lot and improve its delegates
Improve and rename the delegates that AbilitySystemSetupComponent
provides by having them provide a PreviousASC and a NewASC as
parameters.
Also rename delegates:
	SetupWithAbilitySystemCompleted -> OnAbilitySystemSetUp
	PreApplyStartupEffects -> OnAbilitySystemSetUpPreInitialized

Change PreviousPlayerASC to PreviousASC agnostic to Player vs AI. This
makes things easier. And use this PreviousASC to pass into the improved
delegate parameters.

Make a GetAbilitySystemComponent() in AbilitySystemSetupComponent and
use that instead of the owner's. This makes the component more reliable
to work as expected. And makes things easier for the owner to implement
(he just needs to call
UAbilitySystemSetupComponent::GetAbilitySystemComponent() in his
IAbilitySystemInterface::GetAbilitySystemComponent() implementation).

Change ASSGameplayAbility members to generic GameplayAbility types.

Also a lot of general code clean up and some documentation clean up.