Clean up some things and update some documentation.
Test out the StartupAttributeSets and confirm that they are working (we
are now using this in SonicShooter). Update Attribute Set documentation
for it.

Make IAbilitySystemSetupInterface events not pure virtual (make them
optional to override). And update the example code in it.

Delete AIAbilitySystemComponentReplicationMode member because it serves
no purpose. It's not possible to conifgure in BP for the ASC and in CPP
you can just access the AI's ASC and set it instead.