Replace old Startup Abilities, Effects, and Attribute Sets with new GrantSets.
Also create a NativeGameplayTags namespace.
Deleted delegates GiveStartingAbilitiesDelegate and AddStartingAttributeSetsDelegate. This ditches old feature that we found out wasn't useful (allowing game code to cache a memeber pointer to an ability handle). Keeping track of member abilitiy handles in the character to activate them is what was using these, but it's a bad design and clutters code.