Fix early ASC effect application crash due to AttributeSet on rep crash.
Do this by making attribute sets' outer the ASC's owning actor to make engine's getters on attribute sets work as expected.
Getters work now since AttributeSets are now able to access the ASC before the weak ptr is replicated. Prior to UE5, UObject outers weren't replicated with the UObject and instead it just chose the actor channel's actor which is the PlayerState for attribute sets, which is why we didn't encounter this problem before UE5.
This commit also gets rid of unneccesary Rename() call in favor of just using NewObject()'s \"Outer\" parameter.